requirements.txt
streamlit==1.45.0
selenium==4.18.1
pandas==2.2.1
python-dotenv==1.0.1
webdriver-manager==4.0.1
requests==2.31.0
streamlit-authenticator==0.4.2
supabase==1.2.0
PyMuPDF==1.25.5
PyYAML==6.0.2
